final version

+++ src/MemoryManager.h
@@ -103,8 +103,8 @@ static void* CallocWithCheck(size_t x, size_t y)
 
     #else
 
-    #define Malloc(x) malloc(max(x,1))
-    #define Calloc(x,y) calloc(max(x,1),max(y,1))
+    #define Malloc(x) malloc((long unsigned int)max((long unsigned int)x,(long unsigned int)1))
+    #define Calloc(x,y) calloc((long unsigned int)max((long unsigned int)x,(long unsigned int)1),(long unsigned int)max((long unsigned int)y,(long unsigned int)1))
     #define Free(x) free(x)
 
     #endif // ALLOW_ALLOC_ZERO_BYTES
